LPS-56703 Move from portal-impl
