+++ modules/apps/blogs/blogs-web/src/com/liferay/blogs/web/blogs/configuration/BlogsWebConfigurationValues.java
@@ -21,16 +21,13 @@ import com.liferay.portal.kernel.util.GetterUtil;
  */
 public class BlogsWebConfigurationValues {
 
-	public static final String DISPLAY_TEMPLATES_CONFIG =
-		GetterUtil.getString(
-			BlogsWebConfigurationUtil.get("display.templates.config"));
+	public static final String DISPLAY_TEMPLATES_CONFIG = GetterUtil.getString(
+		BlogsWebConfigurationUtil.get("display.templates.config"));
 
-	public static final String ENTRY_CHECK_INTERVAL =
-		GetterUtil.getString(
-			BlogsWebConfigurationUtil.get("entry.check.interval"));
+	public static final String ENTRY_CHECK_INTERVAL = GetterUtil.getString(
+		BlogsWebConfigurationUtil.get("entry.check.interval"));
 
-	public static final String LINKBACK_JOB_INTERVAL =
-		GetterUtil.getString(
-			BlogsWebConfigurationUtil.get("linkback.job.interval"));
+	public static final String LINKBACK_JOB_INTERVAL = GetterUtil.getString(
+		BlogsWebConfigurationUtil.get("linkback.job.interval"));
 
 }
